Move logic from interpreter to interpreter-lib
