+++ libs/interpreter-lib/project.json
@@ -1,6 +1,6 @@
 {
   "name": "interpreter-lib",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "libs/interpreter-lib/src",
   "projectType": "library",
   "targets": {
